[GH-11] - adjusting vb solution tests structure
