+++ tests/NSubstitute.Analyzers.Tests.VisualBasic/NSubstitute.Analyzers.Tests.VisualBasic.csproj
@@ -9,7 +9,7 @@
     <PackageReference Include="NSubstitute" Version="3.1.0" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.3.1" />
     <PackageReference Include="xunit" Version="2.3.1" />
-    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" PrivateAssets="all"/>
+    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" PrivateAssets="all" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\NSubstitute.Analyzers.Shared\NSubstitute.Analyzers.Shared.csproj" />
